feat: move and clean up modals
