+++ manager/src/lib/activities/activities_scripts.ts
@@ -1,0 +1,1 @@
+export function getDefinedActivityTypes() {}
