+++ manager/src/routes/overview/templates/[template_id]/sessions/+page.svelte
@@ -3,7 +3,7 @@
 	import Button from '$components/elements/typography/Button.svelte'; // Verify path
 	import Input from '$components/elements/typography/Input.svelte'; // Verify path
 	import SessionRow from '$components/dashboard/SessionRow.svelte'; // Reuse SessionRow
-	import CreateSessionModal from '$components/elements/CreateSessionModal.svelte';
+	import CreateSessionModal from '$components/elements/modals/CreateSessionModal.svelte';
 	import DataTable from '$components/dashboard/DataTable.svelte';
 	import type { ColumnHeader } from '$components/dashboard/DataTable.svelte';
 	import { page } from '$app/state';
@@ -170,10 +170,9 @@
 		noResultsMessage="You haven't run any sessions yet."
 		bind:searchTerm
 		bind:currentPage
-		totalPages={Math.ceil(filteredSessions.length / 10)}
 	>
 		<svelte:fragment slot="row" let:item>
-			<SessionRow session={item} onActionClick={handleAction} />
+			<SessionRow session={item} />
 		</svelte:fragment>
 	</DataTable>
 </div>
